+++ utils.hpp
@@ -22,19 +22,19 @@ struct TupleHash {
 };
 
 //returns lowest avaliable color in terms of total weight. returns UNDEFINED if no color is avaliable
-int getLowestAvaliableColor(int* totalColorWeights, int* avaliableColors, int colorCount);
+int getLowestAvaliableColor(float* totalColorWeights, int* avaliableColors, int colorCount);
 
 //TODO
 //its a "copy" of getLowestAvaliableColor()
 //only code it if its going to be used
 //returns highest avaliable color in terms of total weight
-int getHighestAvaliableColor(int* totalColorWeights, int* avaliableColors);
+int getHighestAvaliableColor(float* totalColorWeights, int* avaliableColors);
 
 //TODO
 //its a "copy" of getLowestAvaliableColor() and getHighestAvaliableColor()
 //only code it if its going to be used
 //returns array with lowest and highest avaliable color in terms of total weight
-int* getLowestAndHighestAvaliableColor(int* totalColorWeights, int* avaliableColors);
+int* getLowestAndHighestAvaliableColor(float* totalColorWeights, int* avaliableColors);
 
 //changes the color of the node and changes the avaliableColors matrix for consistency
 void updateNodeColor(int* avaliableColors, int nodeCount, int* coloration, int** nodeAjacencyList, int nodeBestColor, int node, int adjacentNodesCount, int colorCount);
